# server/openapi.yaml
openapi: 3.1.0
info:
  title: WealthGenie Public API
  version: 2.0.0
  description: Canonical contract for the supported public Express API. Internal metrics and MCP routes are intentionally excluded.
servers:
  - { url: 'http://localhost:5000', description: Local development }
  - { url: '/', description: Same-origin production gateway }
paths:
...
  /api/tax/compute:
    get:
      operationId: computeTax
      tags: [Tax]
      security: []
      parameters: [{ $ref: '#/components/parameters/Income' }, { $ref: '#/components/parameters/IncomeSource' }, { $ref: '#/components/parameters/FiscalYear' }, { name: age, in: query, required: false, schema: { type: integer, minimum: 18, maximum: 120, description: Required for old-regime calculations and age-dependent deductions. } }, { name: regime, in: query, required: true, schema: { type: string, enum: [new, old] } }]
      responses: { '200': { $ref: '#/components/responses/TaxCalculation' }, '400': { $ref: '#/components/responses/BadRequest' } }
...
  /api/tax/compare:
    get:
      operationId: compareTaxRegimes
      tags: [Tax]
      security: []
      parameters: [{ $ref: '#/components/parameters/Income' }, { $ref: '#/components/parameters/IncomeSource' }, { $ref: '#/components/parameters/FiscalYear' }, { name: age, in: query, required: true, schema: { type: integer, minimum: 18, maximum: 120, description: Required because comparison includes age-sensitive old-regime slabs. } }]
      responses: { '200': { $ref: '#/components/responses/TaxComparison' }, '400': { $ref: '#/components/responses/BadRequest' } }
  /api/tax/post-tax-return:
    post:
      operationId: calculatePostTaxReturn
      tags: [Tax]
      security: []
      x-csrf-protection: cookie-session
      requestBody: { $ref: '#/components/requestBodies/PostTaxReturn' }
      responses: { '200': { $ref: '#/components/responses/Object' }, '400': { $ref: '#/components/responses/BadRequest' } }
  /api/tax/post-tax-return/batch:
    post:
      operationId: calculatePostTaxReturns
      tags: [Tax]
      security: []
      x-csrf-protection: cookie-session
      requestBody: { $ref: '#/components/requestBodies/PostTaxReturnBatch' }
      responses: { '200': { $ref: '#/components/responses/Object' }, '400': { $ref: '#/components/responses/BadRequest' } }
...
components:
  securitySchemes:
    CookieAuth: { type: apiKey, in: cookie, name: wg_session, description: Secure HttpOnly browser session cookie. }
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Explicitly configured non-browser compatibility authentication. }
  parameters:
...
    Income: { name: income, in: query, required: true, schema: { type: number, minimum: 0, maximum: 1000000000 } }
    IncomeSource: { name: incomeSource, in: query, required: true, schema: { type: string, enum: [salary, pension, family_pension, business, other] } }
    FiscalYear: { name: fiscalYear, in: query, required: true, schema: { type: string, enum: [FY2025-26, FY2026-27] } }
  requestBodies:
...
    PostTaxReturn: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PostTaxReturnRequestV2' } } } }
    PostTaxReturnBatch: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PostTaxReturnBatchRequestV2' } } } }
...
  responses:
    Object: { description: Successful structured response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
    TaxCalculation: { description: Authoritative tax calculation, content: { application/json: { schema: { $ref: '#/components/schemas/TaxCalculation' } } } }
    TaxComparison: { description: Authoritative regime comparison, content: { application/json: { schema: { $ref: '#/components/schemas/TaxComparison' } } } }
...
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [error, message, code, request_id]
      properties:
        error: { type: string, description: Backward-compatible summary }
        message: { type: string }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        request_id: { type: [string, 'null'] }
        details: { oneOf: [{ type: array, items: { type: string } }, { type: object, additionalProperties: true }] }
...
    TaxCalculation: { type: object, required: [fiscalYear, policyVersion, inputsUsed], properties: { regime: { type: string, enum: [new, old] }, tax: { type: number }, taxAmount: { type: number }, effectiveRate: { type: number }, fiscalYear: { type: string, pattern: '^FY\\d{4}-\\d{2}$' }, policyVersion: { type: string }, inputsUsed: { type: object, additionalProperties: true }, rulesApplied: { type: array, items: { type: string } }, sourceReferences: { type: array, items: { type: object, additionalProperties: true } } }, additionalProperties: true }
    TaxComparison: { type: object, properties: { old: { $ref: '#/components/schemas/TaxCalculation' }, new: { $ref: '#/components/schemas/TaxCalculation' }, recommendedRegime: { type: string, enum: [old, new] } }, additionalProperties: true }
...
    TaxDeductionInputs: { type: object, additionalProperties: false, description: All deduction identifiers are compatibility input names, not current statutory authority. Positive unclassified other deductions are rejected; zero is accepted only for backward compatibility., properties: { section80C: { type: number, minimum: 0, maximum: 150000, description: Legacy input name; eligibility is fiscal-year and regime specific. }, nps80CCD1B: { type: number, minimum: 0, maximum: 50000, description: Legacy input name; eligibility is fiscal-year and regime specific. }, nps80CCD2: { type: number, minimum: 0 }, basicSalary: { type: number, minimum: 0 }, isGovtEmployee: { type: boolean }, section80D: { type: number, minimum: 0 }, section80D_self: { type: number, minimum: 0 }, section80D_parents: { type: number, minimum: 0 }, parents_senior: { type: boolean }, self_senior: { type: boolean }, hra: { type: number, minimum: 0 }, homeLoanInterest: { type: number, minimum: 0 }, section80EEA: { type: number, minimum: 0 }, savingsInterest: { type: number, minimum: 0 }, section80TTA: { type: number, minimum: 0 }, section80TTB: { type: number, minimum: 0 }, other: { type: number, minimum: 0, maximum: 0, description: Deprecated compatibility input; only zero is accepted. } } }
...
    PostTaxInstrument: { type: object, additionalProperties: false, required: [instrumentType, nominalRate, holdingYears, monthlySIP], properties: { instrumentType: { type: string, maxLength: 50 }, nominalRate: { type: number, minimum: 0, maximum: 1 }, holdingYears: { type: number, minimum: 0.01, maximum: 100 }, monthlySIP: { type: number, minimum: 0, maximum: 100000000 }, acquisitionDate: { type: string, format: date }, redemptionDate: { type: string, format: date }, redemptionChannel: { type: string, enum: [RBI_REDEMPTION, MATURITY_REDEMPTION, SECONDARY_MARKET_SALE] }, acquiredAtOriginalIssue: { type: boolean }, heldContinuously: { type: boolean }, couponRate: { type: number, minimum: 0, maximum: 1 }, annuityFraction: { type: number, minimum: 0, maximum: 1 }, retirementTiming: { type: string, minLength: 2, maxLength: 100 } } }
    PostTaxReturnRequestV2: { allOf: [{ $ref: '#/components/schemas/PostTaxInstrument' }, { type: object, additionalProperties: false, required: [annualIncome, regime, userAge, incomeSource, fiscalYear], properties: { annualIncome: { type: number, minimum: 0, maximum: 1000000000 }, regime: { type: string, enum: [new, old] }, userAge: { type: integer, minimum: 18, maximum: 120 }, incomeSource: { type: string, enum: [salary, pension, family_pension, business, other] }, fiscalYear: { type: string, pattern: '^FY\d{4}-\d{2}$' }, deductions: { $ref: '#/components/schemas/TaxDeductionInputs' }, section112AExemptionUsed: { type: number, minimum: 0, maximum: 125000 } } }], unevaluatedProperties: false }
    PostTaxReturnBatchRequestV2: { type: object, additionalProperties: false, required: [instruments, annualIncome, regime, userAge, incomeSource, inflationRate, fiscalYear], properties: { instruments: { type: array, minItems: 1, maxItems: 50, items: { $ref: '#/components/schemas/PostTaxInstrument' } }, annualIncome: { type: number, minimum: 0, maximum: 1000000000 }, regime: { type: string, enum: [new, old] }, userAge: { type: integer, minimum: 18, maximum: 120 }, incomeSource: { type: string, enum: [salary, pension, family_pension, business, other] }, inflationRate: { type: number, minimum: 0, maximum: 1 }, fiscalYear: { type: string, pattern: '^FY\d{4}-\d{2}$' }, deductions: { $ref: '#/components/schemas/TaxDeductionInputs' }, section112AExemptionUsed: { type: number, minimum: 0, maximum: 125000 } } }
...